regorganize view's helpers
